web/requirements.txt
django-cms>=3.5,<3.6
djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0
django-webpack-loader==0.6.0
django-mathfilters==0.1.2

djangocms-text-ckeditor>=3.6.0
djangocms-link>=2.1
djangocms-style>=2.0
djangocms-googlemap>=1.1
djangocms-snippet>=2.0
djangocms-video>=2.0
djangocms-column>=1.7
djangocms-file>=2.0,<3.0
djangocms-picture>=2.0,<3.0
easy_thumbnails
django-filer>=1.3
Django<2.0
pytz
django-classy-tags>=0.7
html5lib>=0.999999,<0.99999999
Pillow>=3.0
django-sekizai>=0.9
six

psycopg2==2.7.5
gunicorn==19.3.0
django-widget-tweaks==1.4.2

django-parler==1.9.2
djangorestframework==3.8.2